requirements.txt
flask==2.3.2
flask-cors==4.0.0
pillow==10.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0